Remove comments, fix logging, rename Request generator, fix memory caching condition
