+++ src/main/java/io/cryostat/recordings/ActiveRecordings.java
@@ -29,8 +29,8 @@ import java.util.UUID;
 import io.cryostat.ConfigProperties;
 import io.cryostat.libcryostat.templates.Template;
 import io.cryostat.libcryostat.templates.TemplateType;
-import io.cryostat.recordings.ArchiveRequestGenerator.ArchiveRequest;
-import io.cryostat.recordings.ArchiveRequestGenerator.GrafanaActiveUploadRequest;
+import io.cryostat.recordings.LongRunningRequestGenerator.ArchiveRequest;
+import io.cryostat.recordings.LongRunningRequestGenerator.GrafanaActiveUploadRequest;
 import io.cryostat.recordings.RecordingHelper.RecordingOptions;
 import io.cryostat.recordings.RecordingHelper.RecordingReplace;
 import io.cryostat.targets.Target;
@@ -67,7 +67,7 @@ public class ActiveRecordings {
 
     @Inject ObjectMapper mapper;
     @Inject RecordingHelper recordingHelper;
-    @Inject ArchiveRequestGenerator generator;
+    @Inject LongRunningRequestGenerator generator;
     @Inject EventBus bus;
     @Inject Logger logger;
 
@@ -130,31 +130,16 @@ public class ActiveRecordings {
                         .atMost(connectionFailedTimeout);
                 return null;
             case "save":
-                // FIXME this operation might take a long time to complete, depending on the
-                // amount of JFR data in the target and the speed of the connection between the
-                // target and Cryostat. We should not make the client wait until this operation
-                // completes before sending a response - it should be async. Here we should just
-                // return an Accepted response, and if a failure occurs that should be indicated
-                // as a websocket notification.
-
-                /*
-                 * Desired workflow:
-                 *   Client sends a PATCH request to Cryostat
-                 *   Cryostat receives the PATCH and checks that the specified active recording exists and that the target JVM is reachable (ex. try to open a connection and do something relatively lightweight like compute its JVM ID). If this check succeeds respond to the PATCH with 202, if it fails respond with a 404 (recording not found) or 502 (target not reachable) etc.
-                 *   In the background, Cryostat creates the S3 file upload request, opens a target connection, pipes the bytes, etc. - same as steps 2-5 above
-                 *   Cryostat emits a WebSocket notification, either indicating task successful completion or task failure.
-                 */
-                logger.info("Ceating request");
                 ArchiveRequest request =
                         new ArchiveRequest(UUID.randomUUID().toString(), activeRecording);
-                logger.info(
+                logger.tracev(
                         "Request created: ("
                                 + request.getId()
                                 + ", "
                                 + request.recording().name
                                 + ")");
                 response.endHandler(
-                        (e) -> bus.publish(ArchiveRequestGenerator.ARCHIVE_ADDRESS, request));
+                        (e) -> bus.publish(LongRunningRequestGenerator.ARCHIVE_ADDRESS, request));
                 return request.getId();
             default:
                 throw new BadRequestException(body);
@@ -262,7 +247,7 @@ public class ActiveRecordings {
                         + request.getTargetId()
                         + ")");
         response.endHandler(
-                (e) -> bus.publish(ArchiveRequestGenerator.GRAFANA_ACTIVE_ADDRESS, request));
+                (e) -> bus.publish(LongRunningRequestGenerator.GRAFANA_ACTIVE_ADDRESS, request));
         return request.getId();
     }
